#2308 Fix r26358 for clang compiler
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@26368 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/src/utilities/BoxCollection.cpp
@@ -311,7 +311,7 @@ bool BoxCollection<DIM>::IsBoxInDomain(c_vector<unsigned, DIM> gridIndices)
     {
         if (!mIsDomainPeriodic(dim))
         {
-            if (gridIndices(dim) >= (int)mNumBoxesEachDirection(dim))
+            if (gridIndices(dim) >= mNumBoxesEachDirection(dim))
             {
                 is_in_domain = false;
             }
